custom hook, handling for status codes
